Change float to double in radial function

+++ src/cpp/math/radial_function.cpp
@@ -23,11 +23,11 @@ RadialFunction::RadialFunction(unsigned n, unsigned l) : m_N{n}, m_L{l}, m_R{0} 
   updateValue();
 }
 
-float RadialFunction::operator()() const {
+double RadialFunction::operator()() const {
   return m_Value;
 }
 
-float RadialFunction::operator()(float r) {
+double RadialFunction::operator()(double r) {
   if (r != m_R) {
     setR(r);
   }
@@ -48,7 +48,7 @@ void RadialFunction::setL(unsigned l) {
   updateValue();
 }
 
-void RadialFunction::setR(float r) {
+void RadialFunction::setR(double r) {
   assert(r >= 0);
   m_R = r;
   updateValue();
@@ -62,21 +62,21 @@ unsigned RadialFunction::getL() const {
   return m_L;
 }
 
-float RadialFunction::getR() const {
+double RadialFunction::getR() const {
   return m_R;
 }
 
 void RadialFunction::updateValue() {
-  const float a = (2 * m_R) / (m_N * BOHR_IN_PM);
-  m_Value = m_normFactor * exp(-a / 2) * powf(a, m_L) *
-            boost::math::laguerre<float>(m_N - m_L - 1, 2 * m_L + 1, a);
+  const double a = (2 * m_R) / (m_N * BOHR_IN_PM);
+  m_Value = m_normFactor * exp(-a / 2) * pow(a, m_L) *
+            boost::math::laguerre<double>(m_N - m_L - 1, 2 * m_L + 1, a);
 }
 
 void RadialFunction::updateNormFactor() {
-  const float a = powf(2 / (m_N * BOHR_IN_PM), 3);
-  const float b = factorial<float>(m_N - m_L - 1) / (2 * m_N * factorial<float>(m_N + m_L));
+  const double a = pow(2 / (m_N * BOHR_IN_PM), 3);
+  const double b = factorial<double>(m_N - m_L - 1) / (2 * m_N * factorial<double>(m_N + m_L));
   // idk from where that 384.948 comes from, but it manages to normalize the function, so...
-  m_normFactor = 384.948f * sqrtf(a * b);
+  m_normFactor = 384.948 * sqrt(a * b);
 }
 
 }  // namespace math
